requirements.txt
Flask>=3.0.0
python-telegram-bot>=21.0
pdf2docx>=0.5.8
python-docx>=1.1.0
Pillow>=10.3.0
pytesseract>=0.3.10
gunicorn>=22.0.0
